allow iq-by-action structs to have null fields
I don't have a conclusive answer for why these fields can be null, but my best guess is that auth tokens associated with a user will fill them in,
but tokens that aren't associated with a user (client tokens) don't have any reasonable data for them.

While it certainly is inconvenient to have structs with all fields as Option<>, there's no way to discern between user tokens and client tokens.